pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "aiosqlite>=0.21.0",
    "discord-py>=2.6.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
